requirements.txt
Flask==1.0.2
flask-swagger-ui==3.20.9
flake8==3.7.7
apispec-webframeworks==0.4.0
flask-apispec==0.8.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
SQLAlchemy-Utils==0.34.1
Flask-Login==0.4.1
mysqlclient==1.4.2.post1
passlib==1.7.1
